Sources/GhostyApp/Resources/Backend/pyproject.toml
[project]
name = "ghosty-backend"
version = "0.1.0"
description = "Python backend for Ghosty assistant"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "modal>=0.63.0",
    "langchain>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "pillow>=10.0.0",
    "requests>=2.30.0",
]

[tool.uv]
managed = true
